refactor: Move DynamicTagValues to File
