+++ libwild/src/grouping.rs
@@ -306,7 +306,7 @@ impl<'data> SequencedInputObject<'data> {
     }
 
     pub(crate) fn is_dynamic(&self) -> bool {
-        self.parsed.dynamic_tag_values.is_some()
+        self.parsed.object.dynamic_tag_values.is_some()
     }
 }
